Fix various bugs with links, ids, and the table of contents
Generating links is tricky. We need to make sure to URI-escape anchors in the
link.

Make sure to use user-provided content for the link text (inside the <a> tag)
but modify it as appropriate (for example to remove a leading '#' on an anchor
link, or a leading 'doc:' on a module link).

When generating IDs, we need to use the raw Pod but then HTML-escape that,
rather than just using the text contents of the things we're generating an ID
for. This makes it much easier for people to write links to headings that
contain Pod. They can write L<#Heading with C<code>> rather than having to
realize that the C<> tag is stripped out for the ID.

I also fixed how we generate the TOC. Rather than generating Pod as a string
and EVALing it, I just make the appropriate Pod objects (Items containing L<>
FormattingCode objects) and pass that to pod-to-html. Trying to generate Pod
as a string was very fragile. If we just write "L<$text>" then a piece of text
like "gt - >" breaks the Pod.